+++ src/CommandSystem/Handler.h
@@ -27,6 +27,13 @@ using StopNowFuncType = std::function<void()>;
 using IsWritingFuncType = std::function<bool()>;
 using GetJobIdFuncType = std::function<std::string()>;
 
+struct CmdResponse {
+  Log::Severity LogLevel;
+  int StatusCode{0};
+  bool SendResponse;
+  std::function<std::string()> MessageString;
+};
+
 class HandlerBase {
 public:
   HandlerBase() = default;
@@ -52,7 +59,8 @@ public:
   Handler(std::string const &ServiceIdentifier,
           Kafka::BrokerSettings const &Settings, uri::URI JobPoolUri,
           uri::URI CommandTopicUri);
-  Handler(std::string ServiceIdentifier,
+  Handler(std::string const &ServiceIdentifier,
+          Kafka::BrokerSettings const &Settings, uri::URI CommandTopicUri,
           std::unique_ptr<JobListener> JobConsumer,
           std::unique_ptr<CommandListener> CommandConsumer,
           std::unique_ptr<FeedbackProducerBase> Response);
@@ -70,18 +78,70 @@ public:
                                    std::string const &ErrorMessage) override;
 
   void loopFunction() override;
+  bool isUsingAlternativeTopic() const { return UsingAltTopic; }
+
+protected:
+  /// \brief Initiates writing.
+  ///
+  /// \param StartJob The start message as a StartMessage struct.
+  /// \param IsJobPoolCommand Flag to indicate if the command comes from the job
+  /// pool or the command topic.
+  /// \return Metadata about the success/failure after processing the command.
+  CmdResponse startWriting(StartMessage &StartJob, bool IsJobPoolCommand);
+
+  /// \brief Stops writing.
+  ///
+  /// \param StopJob The stop message as a StopMessage struct.
+  /// \return Metadata about the success/failure after processing the command.
+  CmdResponse stopWriting(StopMessage &StopJob);
 
 private:
+  /// \brief Parse a command message and route it to appropriate handling
+  /// method.
+  ///
+  /// \param CommandMsg Kafka message.
+  /// \param IsJobPoolCommand Flag to indicate if the command comes from the job
+  /// pool or the command topic.
   void handleCommand(FileWriter::Msg CommandMsg, bool IsJobPoolCommand);
+
+  /// \brief Handle start command.
+  ///
+  /// \param CommandMsg Kafka message.
+  /// \param IsJobPoolCommand Flag to indicate if the command comes from the job
+  /// pool or the command topic.
   void handleStartCommand(FileWriter::Msg CommandMsg, bool IsJobPoolCommand);
+
+  /// \brief Validate command and start writing.
+  ///
+  /// \param CommandMsg Kafka message.
+  /// \param StartJob Returns the parsed start message as a StartMessage struct.
+  /// \param IsJobPoolCommand Flag to indicate if the command comes from the job
+  /// pool or the command topic.
+  /// \return Metadata about the success/failure after processing the command.
+  CmdResponse startWritingProcess(const FileWriter::Msg &CommandMsg,
+                                  StartMessage &StartJob,
+                                  bool IsJobPoolCommand);
+
+  /// \brief Handle stop command.
+  ///
+  /// \param CommandMsg Kafka message.
   void handleStopCommand(FileWriter::Msg CommandMsg);
+
+  /// \brief Validate command and start writing.
+  ///
+  /// \param CommandMsg Kafka message.
+  /// \param StopJob Returns the parsed stop message as a StopMessage struct.
+  /// \return Metadata about the success/failure after processing the command.
+  CmdResponse stopWritingProcess(const FileWriter::Msg &CommandMsg,
+                                 StopMessage &StopJob);
+
   std::string const ServiceId;
   std::string NexusStructure;
   StartFuncType DoStart{[](auto) {
     throw std::runtime_error("DoStart(): Not set/implemented.");
   }};
   StopTimeFuncType DoSetStopTime{[](auto) {
-    throw std::runtime_error("DosetStopTime(): Not set/implemented.");
+    throw std::runtime_error("DoSetStopTime(): Not set/implemented.");
   }};
   StopNowFuncType DoStopNow{
       []() { throw std::runtime_error("DoStopNow(): Not set/implemented."); }};
@@ -91,6 +151,10 @@ private:
   GetJobIdFuncType GetJobId{[]() -> std::string {
     throw std::runtime_error("GetJobId(): Not set/implemented.");
   }};
+
+  /// \brief Switch to an alternative command topic.
+  void switchCommandTopic(std::string_view ControlTopic,
+                          time_point const StartTime);
 
   /// \brief Revert to the default command topic if an alternative topic
   /// has been configured.